back-end/requirements.txt
Flask>=2.0
psycopg2-binary >=1.0
python-dotenv
